#62 refactor: 타입 정의를 schemas.ts로 통합하고 불필요한 타입 파일 제거
