+++ src/entities/student/model/types.ts
@@ -1,5 +1,5 @@
 import type {z} from 'zod';
-import type {ApiResponse} from '@/shared/model/type';
+import type {ApiResponse} from '@/shared/model/types';
 import {
   progressStatusSchema,
   studentProgressSchema,
