+++ src/features/assignment/filter-assignment/lib/useAssignmentList.ts
@@ -1,14 +1,14 @@
 import {useQuery} from '@tanstack/react-query';
 import {assignmentQueries} from '@/entities/assignment/api/assignmentQueries';
-import type {Assignment} from '@/entities/assignment/model/types';
+import type {TAssignment} from '@/entities/assignment/model/schemas';
 
 // 중복 제거
-const unique = (list: Assignment[]) =>
+const unique = (list: TAssignment[]) =>
   Array.from(new Map(list.map((a) => [a.id, a])).values());
 
 export const useAssignmentList = (
   selectedCourseId: number | null
-): Assignment[] => {
+): TAssignment[] => {
   const {data: allAssignments} = useQuery(
     assignmentQueries.getAllAssignments()
   );
