+++ src/pages/unit-editor/ui/UnitAssignmentList.tsx
@@ -1,5 +1,5 @@
 import ListRow from '@/shared/ui/list-row/ListRow';
-import type {Assignment} from '@/entities/assignment/model/types';
+import type {TAssignment} from '@/entities/assignment/model/schemas';
 import {useState} from 'react';
 import DragAndDropIcon from '@/assets/svg/dragAndDropIcon.svg?react';
 import {
@@ -24,9 +24,9 @@ import {CSS} from '@dnd-kit/utilities';
 export const UnitAssignmentList = ({
   assignmentList,
 }: {
-  assignmentList: Assignment[];
+  assignmentList: TAssignment[];
 }) => {
-  const [assignments, setAssignments] = useState<Assignment[]>(assignmentList);
+  const [assignments, setAssignments] = useState<TAssignment[]>(assignmentList);
 
   // 과제 인덱스 찾기
   const getAssignmentIndex = (id: UniqueIdentifier) => {
@@ -80,7 +80,7 @@ export const UnitAssignmentList = ({
 };
 
 // 드래그 가능한 과제 아이템 컴포넌트
-const DraggableAssignmentItem = ({id, title}: Assignment) => {
+const DraggableAssignmentItem = ({id, title}: TAssignment) => {
   const {attributes, listeners, setNodeRef, transform, transition, isDragging} =
     useSortable({id});
