+++ src/shared/lib/course.ts
@@ -1,5 +1,5 @@
-import type {Unit} from '@/entities/course/model/types';
-import type {SemesterCode} from '@/shared/model/type';
+import type {TUnit} from '@/entities/unit/model/schemas';
+import type {SemesterCode} from '@/shared/model/types';
 
 const SEMESTER_MAP: Record<SemesterCode, '1' | '2' | '여름' | '겨울'> = {
   FIRST: '1',
@@ -50,7 +50,7 @@ export const formatCourseTerm = (
 };
 
 // 총 과제 수 계산
-export const getTotalAssignmentCount = (units: Unit[]): number => {
+export const getTotalAssignmentCount = (units: TUnit[]): number => {
   return units.reduce((acc, unit) => acc + unit.assignmentCount, 0);
 };
